Download cra file
