+++ django/api/apps.py
@@ -1,7 +1,5 @@
 from django.apps import AppConfig
-from django.contrib.admin import AdminSite
 from django.contrib.admin.apps import AdminConfig
-from .views import download_file
 
 
 class ApiConfig(AppConfig):
@@ -11,17 +9,5 @@ class ApiConfig(AppConfig):
         import api.signals
 
 
-class ITVRAdminSite(AdminSite):
-    def get_urls(self):
-        from django.urls import path
-
-        print("GETTING CUSTOM URLS")
-
-        urls = super().get_urls()
-        urls = [path("cra-download", self.admin_view(download_file))] + urls
-
-        return urls
-
-
 class ITVRAdminConfig(AdminConfig):
-    default_site = "api.apps.ITVRAdminSite"
+    default_site = "api.site.ITVRAdminSite"
